Fix: Lotto.java 추상화
수동 로또, 자동 로또, 당첨 로또는 로또라는 공통점을 갖고 있기에 추상화로 변경
